build(nix): use `crane` and `bun2nix` in flake
Instead of installing and running bun and the rust toolchain as
packages, we make use of `crane` and `bun2nix` to create fixed output
derivations which can be cached entirely in the nix store.

This is part 1 of the implementation for
https://github.com/CompeyDev/ssh-portfolio/issues/6.

Commit includes some minor refactoring and reformatting of code too.